+++ cmd/policy.go
@@ -28,6 +28,7 @@ func newPolicyCmd() *cobra.Command {
 
 	cmd.AddCommand(newPolicyDisableCmd())
 	cmd.AddCommand(newPolicyEnableCmd())
+	cmd.AddCommand(newPolicyGroupsCmd())
 	cmd.AddCommand(newPolicyLsCmd())
 	cmd.AddCommand(newPolicyNewCmd())
 	cmd.AddCommand(newPolicyPublishCmd())
