+++ tests/integration/policy/policy_test.go
@@ -38,11 +38,11 @@ func TestPolicy(t *testing.T) {
 	e.RunCommand("pulumi", "policy", "publish", orgName)
 
 	// Check the policy ls commands.
-	packsOutput, _ := e.RunCommand("pulumi", "policy", "ls", "packs", "--json")
+	packsOutput, _ := e.RunCommand("pulumi", "policy", "ls", "--json")
 	var packs []policyPacksJSON
 	assertJSON(e, packsOutput, &packs)
 
-	groupsOutput, _ := e.RunCommand("pulumi", "policy", "ls", "groups", "--json")
+	groupsOutput, _ := e.RunCommand("pulumi", "policy", "groups", "ls", "--json")
 	var groups []policyGroupsJSON
 	assertJSON(e, groupsOutput, &groups)
